requirements.txt
flask
tensorflow
numpy
pillow
h5py
opencv-python